requirements_for_venv.txt
anyio==4.8.0
asgiref==3.8.1
asyncpg==0.30.0
certifi==2025.1.31
Django==5.1.7
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
pillow==11.1.0
psycopg2-binary==2.9.10
python-telegram-bot==21.11.1
sniffio==1.3.1
SQLAlchemy==2.0.38
sqlparse==0.5.3
typing_extensions==4.12.2
